Set petList to formatted pet name

+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/myPet/petFeedScheduler/PetFeedScheduleListItem.kt
@@ -2,9 +2,9 @@ package com.sju18001.petmanagement.ui.myPet.petFeedScheduler
 
 import java.time.LocalTime
 
-class PetFeedScheduleListItem(_feedTime: LocalTime, _petList: ArrayList<Int>, _memo: String, _isTurnedOn: Boolean) {
+class PetFeedScheduleListItem(_feedTime: LocalTime, _petList: ArrayList<Long>, _memo: String, _isTurnedOn: Boolean) {
     var feedTime: LocalTime = _feedTime
-    var petList: ArrayList<Int> = _petList
+    var petList: ArrayList<Long> = _petList
     var memo: String = _memo
     var isTurnedOn: Boolean = _isTurnedOn
 }
